+++ DocCodeSamples.Tests/LocalizationSettingsSamples.cs
@@ -76,3 +76,38 @@ public class SelectedLocaleChangedExample : MonoBehaviour
     }
 }
 #endregion
+
+#region force-update
+
+public class ForceUpdateExample : MonoBehaviour
+{
+    public LocalizedString myString = new LocalizedString("MyTable", "MyEntry");
+
+    void OnEnable()
+    {
+        myString.StringChanged += UpdateText;
+    }
+
+    void OnDisable()
+    {
+        myString.StringChanged -= UpdateText;
+    }
+
+    void UpdateText(string text)
+    {
+        Debug.Log("The value is " + text);
+    }
+
+    // This method is called to change the value in the string table.
+    [ContextMenu("Change Source Value")]
+    public void ChangeSourceValue()
+    {
+        // Get the table and entry, and update the value. 
+        var tableEntry = LocalizationSettings.StringDatabase.GetTableEntry(myString.TableReference, myString.TableEntryReference);
+        tableEntry.Entry.Value = "New Value";
+
+        // Force a refresh to update everything, including any uses of the changed entry.
+        LocalizationSettings.Instance.ForceRefresh();
+    }
+}
+#endregion