+++ Editor/UI/Smart Format/PersistentVariablePropertyDrawer.cs
@@ -11,12 +11,13 @@ namespace UnityEditor.Localization.UI
         public override VisualElement CreatePropertyGUI(SerializedProperty property)
         {
             var value = property.FindPropertyRelative("m_Value");
-            return new PropertyField(value, ManagedReferenceUtility.GetDisplayName(property.managedReferenceFullTypename).text);
+            var label = property.propertyType == SerializedPropertyType.ManagedReference ? ManagedReferenceUtility.GetDisplayName(property.managedReferenceFullTypename).text : property.displayName;
+            return new PropertyField(value, label);
         }
 
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
-            if (string.IsNullOrEmpty(label.text) && property.propertyType == SerializedPropertyType.ManagedReference)
+            if (property.propertyType == SerializedPropertyType.ManagedReference)
                 label = ManagedReferenceUtility.GetDisplayName(property.managedReferenceFullTypename);
             EditorGUI.PropertyField(position, property.FindPropertyRelative("m_Value"), label, true);
         }